+++ example/bonhomme/model/bonhomme.dart
@@ -1,0 +1,13 @@
+part of bonhomme;
+
+class Bonhomme extends MovablePiece {
+  Bonhomme([int id]): super(id) {
+    randomInit();
+    shape = PieceShape.IMG;
+    width = 60;
+    height = 60;
+    x = 0;
+    y = 0;
+    imgId = 'bonhomme';
+  }
+}